+++ resources/[dg]/dg-misc/typescript/src/client/controllers/events.ts
@@ -2,7 +2,6 @@ import { handleArenaModuleResourceStop } from 'modules/arena/service.arena';
 import { handleHudModuleResourceStop } from 'modules/hud/service.hud';
 import { handleObjectManagerModuleResourceStop } from 'modules/objectManager/service.objectmanager';
 import { handlePropattachModuleResourceStop } from 'modules/propattach/service.propattach';
-import { handleRentalModuleResourceStop } from 'modules/rental/service.rental';
 import { handleAnimationsServiceResourceStop } from 'services/animations';
 
 on('onResourceStop', (resourceName: string) => {
@@ -12,7 +11,6 @@ on('onResourceStop', (resourceName: string) => {
   handleHudModuleResourceStop();
   handleObjectManagerModuleResourceStop();
   handlePropattachModuleResourceStop();
-  handleRentalModuleResourceStop();
 
   handleAnimationsServiceResourceStop();
 });